+++ ires.Domain/Contracts/ISurveyService.cs
@@ -8,10 +8,10 @@ namespace ires.Domain.Contracts
         public Task<SurveyViewModel> Create(SurveyRequestDto requestDto);
         public Task<bool> Update(SurveyRequestDto requestDto);
         public Task<bool> UpdateStatus(long ID, SurveyStatus status);
-        public Task<ICollection<SurveyViewModel>> GetSurveys(long companyID, string search);
+        public Task<ICollection<SurveyViewModel>> GetSurveys(int companyID, string search);
         public Task<SurveyViewModel> GetByID(long id);
-        public Task<int> CountPending(long companyID);
-        public Task<int> CountCompleted(long companyID);
+        public Task<int> CountPending(int companyID);
+        public Task<int> CountCompleted(int companyID);
 
     }
 }
